# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Node.js + Mysql",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/cities/cities/{name}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/cities/cities/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "code": {
                  "example": "any"
                },
                "country_id": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/booking/addBooking": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "senderId": {
                  "example": "any"
                },
                "receiverId": {
                  "example": "any"
                },
                "fromCountryId": {
                  "example": "any"
                },
                "fromCityId": {
                  "example": "any"
                },
                "fromZipCode": {
                  "example": "any"
                },
                "fromAddress": {
                  "example": "any"
                },
                "toCountryId": {
                  "example": "any"
                },
                "toCityId": {
                  "example": "any"
                },
                "toZipCode": {
                  "example": "any"
                },
                "toAddress": {
                  "example": "any"
                },
                "amount": {
                  "example": "any"
                },
                "discountAmount": {
                  "example": "any"
                },
                "totalAmount": {
                  "example": "any"
                },
                "receivedAmount": {
                  "example": "any"
                },
                "payableAmount": {
                  "example": "any"
                },
                "quantity": {
                  "example": "any"
                },
                "weight": {
                  "example": "any"
                },
                "bookingDate": {
                  "example": "any"
                },
                "isDocument": {
                  "example": "any"
                },
                "bookingStatus": {
                  "example": "any"
                },
                "paymentStatus": {
                  "example": "any"
                },
                "shipById": {
                  "example": "any"
                },
                "bookingDetail": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
    "/booking/getAllWithSenderId": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "pageNo": {
                  "example": "any"
                },
                "pageSize": {
                  "example": "any"
                },
                "senderId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/booking/singleBooking/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/booking/BookingsBySenderId/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/booking/booking/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "senderId": {
                  "example": "any"
                },
                "receiverId": {
                  "example": "any"
                },
                "fromCountryId": {
                  "example": "any"
                },
                "fromCityId": {
                  "example": "any"
                },
                "fromZipCode": {
                  "example": "any"
                },
                "fromAddress": {
                  "example": "any"
                },
                "toCountryId": {
                  "example": "any"
                },
                "toCityId": {
                  "example": "any"
                },
                "toZipCode": {
                  "example": "any"
                },
                "toAddress": {
                  "example": "any"
                },
                "amount": {
                  "example": "any"
                },
                "discountAmount": {
                  "example": "any"
                },
                "totalAmount": {
                  "example": "any"
                },
                "receivedAmount": {
                  "example": "any"
                },
                "payableAmount": {
                  "example": "any"
                },
                "quantity": {
                  "example": "any"
                },
                "weight": {
                  "example": "any"
                },
                "bookingDate": {
                  "example": "any"
                },
                "isDocument": {
                  "example": "any"
                },
                "bookingStatus": {
                  "example": "any"
                },
                "paymentStatus": {
                  "example": "any"
                },
                "shipById": {
                  "example": "any"
                },
                "bookingDetail": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  }
}